+++ src/otx/backend/__init__.py
@@ -1,0 +1,4 @@
+"""OTX backends."""
+
+# Copyright (C) 2025 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
