+++ src/otx/backend/native/__init__.py
@@ -1,0 +1,4 @@
+"""Native backend."""
+
+# Copyright (C) 2025 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
